arm64: change if+return to just return, where possible
If a switch is followed by "return false", it is possible to just "return a
== b" instead of "if a == b { return true }" within the cases of a
switch. This change produces shorter assembly output.

There are also other minor changes that either makes the code shorter or
uses switch/case instead of more verbose if checks.

There is no change in behavior.

Change-Id: Ifc4f33ddc5f12e53d7e6f04ac1e054c1e589968b